requirements.txt
discord.py
python-dotenv
Flask
beautifulsoup4
requests